+++ server/src/api/controllers/RedirectShortUrl.controller.ts
@@ -1,12 +1,14 @@
 import IShortUrl from "../models/interfaces/IShortUrl";
 import { ShortUrls } from "../models/shortURL.model";
-import IErrorMessage from "./interfaces/IErrorMessage";
 import { Response } from "express";
 import IRequest from "./interfaces/IRequest";
 import IReqParams from "./interfaces/IReqParams";
+import IMessage from "./interfaces/IMessage";
+import IResponse from "./interfaces/IResponse";
 
 const RedirectShortUrl = async (req: IRequest, res: Response) => {
-  const errorMessage: IErrorMessage = { code: -1, message: "" };
+  const msg: IMessage = { isSuccessful: false, message: "", info: null };
+  const result: IResponse = { code: -1, data: msg };
 
   try {
     const { id }: IReqParams = req.params;
@@ -18,20 +20,27 @@ const RedirectShortUrl = async (req: IRequest, res: Response) => {
     };
 
     if (data.long) {
-      return res.redirect(301, data.long);
+      msg.isSuccessful = true;
+      result.code = 301;
+      msg.message = `Successfully Redirected short URL to its destination`;
+      msg.info = data.long;
+      return res.redirect(result.code, msg.info);
     } else {
-      errorMessage.code = 404;
-      errorMessage.message = `Could not find the long link for this short link ${id}`;
-
-      return res.json({ error: errorMessage });
+      msg.isSuccessful = false;
+      result.code = 404;
+      msg.message = `404 Not Found`;
+      msg.info = `Could not find the long link for this short link ${id}`;
+      return res.status(result.code).json(result.data);
     }
   } catch (error) {
     if (error instanceof Error) {
-      errorMessage.code = 500;
-      errorMessage.message = error.message;
+      msg.isSuccessful = false;
+      result.code = 500;
+      msg.message = error.name;
+      msg.info = error.message;
     }
 
-    return res.json({ error: errorMessage });
+    return res.status(result.code).json(result.data);
   }
 };